+++ test/runtests.jl
@@ -12,5 +12,6 @@ const TEST_ATOL = 1e-6
 @testset "EnergyModelsBase" begin
     include("user_interface.jl")
     include("nodes.jl")
+    include("test_examples.jl")
     include("test_utils.jl")
 end